forge fmt

+++ src/skills/VoteSkill.sol
@@ -13,14 +13,11 @@ import "../proposals/IProposals.sol";
 import "../proposals/AminalProposals.sol";
 
 contract VoteSkill is ISkill, AminalProposals {
-
     uint256 public LoveQuorum = 80;
     uint256 public LoveQuorumDecayPerWeek = 10;
     uint256 public LoveRequiredMajority = 50;
 
-
-
-    constructor(address _aminals) AminalProposals(_aminals){
+    constructor(address _aminals) AminalProposals(_aminals) {
         aminals = _aminals;
     }
 
@@ -32,10 +29,18 @@ contract VoteSkill is ISkill, AminalProposals {
 
         (uint256 proposalId, bool vote) = abi.decode(data, (uint256, bool));
 
-        return _vote(aminalId, sender, proposalId, vote, Aminals(aminals).getAminalLoveTotal(aminalId), LoveQuorum, LoveRequiredMajority);
+        return _vote(
+            aminalId,
+            sender,
+            proposalId,
+            vote,
+            Aminals(aminals).getAminalLoveTotal(aminalId),
+            LoveQuorum,
+            LoveRequiredMajority
+        );
     }
 
-        // function _vote(uint256 aminalID, address sender, uint256 proposalId, bool yesNo, uint256 membersLength, uint256 quorum, uint256 requiredMajority) internal returns (uint256 squeak) {
+    // function _vote(uint256 aminalID, address sender, uint256 proposalId, bool yesNo, uint256 membersLength, uint256 quorum, uint256 requiredMajority) internal returns (uint256 squeak) {
 
     // Getters
     function getSkillData(uint256 proposalId, bool vote) public pure returns (bytes memory data) {
@@ -53,21 +58,28 @@ contract VoteSkill is ISkill, AminalProposals {
         uint256 requiredMajority
     );
 
-
     // Internal functions
 
     // THIS IS A MERITOCRACY BASED ON LOVE THAT AMINAL HAS FOR MSG.SENDER
 
-    function _vote(uint256 aminalID, address sender, uint256 proposalId, bool yesNo, uint256 membersLength, uint256 quorum, uint256 requiredMajority) internal returns (uint256 squeak) {
+    function _vote(
+        uint256 aminalID,
+        address sender,
+        uint256 proposalId,
+        bool yesNo,
+        uint256 membersLength,
+        uint256 quorum,
+        uint256 requiredMajority
+    ) internal returns (uint256 squeak) {
         // replace with squeak calc based on proposal
         squeak = 2;
 
         // TODO: vote calc and execute when successful
-        LoveProposal storage proposal = loveProposals[proposalId]; 
+        LoveProposal storage proposal = loveProposals[proposalId];
         require(proposal.closed == 0);
 
         // uint love = aminals[aminalID].lovePerUser[msg.sender];
-        uint love = Aminals(aminals).getAminalLoveByIdByUser(aminalID, msg.sender);
+        uint256 love = Aminals(aminals).getAminalLoveByIdByUser(aminalID, msg.sender);
 
         // first vote
         if (loveVotes[proposalId][msg.sender] == 0) {
@@ -80,7 +92,7 @@ contract VoteSkill is ISkill, AminalProposals {
             }
             emit LoveVoted(proposalId, aminalID, yesNo, proposal.votedYes, proposal.votedNo);
 
-        // Changing Yes to No
+            // Changing Yes to No
         } else if (loveVotes[proposalId][msg.sender] == 1 && !yesNo && proposal.votedYes > 0) {
             proposal.votedYes--;
             proposal.votedNo++;
@@ -94,12 +106,14 @@ contract VoteSkill is ISkill, AminalProposals {
             emit LoveVoted(proposalId, aminalID, yesNo, proposal.votedYes, proposal.votedNo);
         }
 
-       uint256 voteCount = proposal.votedYes + proposal.votedNo;
+        uint256 voteCount = proposal.votedYes + proposal.votedNo;
         if (voteCount * 100 >= quorum * membersLength) {
             uint256 yesPercent = proposal.votedYes * 100 / voteCount;
             proposal.pass = yesPercent >= requiredMajority;
-            emit LoveVoteResult(proposalId, proposal.pass, voteCount, quorum, membersLength, yesPercent, requiredMajority);
-        } 
+            emit LoveVoteResult(
+                proposalId, proposal.pass, voteCount, quorum, membersLength, yesPercent, requiredMajority
+            );
+        }
 
         return squeak;
     }
