+++ src/utils/VisualsAuction.sol
@@ -180,7 +180,7 @@ contract VisualsAuction is IAminalStructs {
             visualVoted[msg.sender][auctionId][category] < totallove, "Already consumed all of your love with votes"
         );
 
-         console.log("********** a vote has been casted on category: ", category, " /  index: ", i);
+        console.log("********** a vote has been casted on category: ", category, " /  index: ", i);
         // console.log(" == with weight = ", totallove, " .  on auctionId = ", auctionId);
 
         auction.visualIdVotes[i][category] += (totallove);
@@ -205,40 +205,45 @@ contract VisualsAuction is IAminalStructs {
         auction.visualNoVotes[visualId][category] += (totallove);
 
         console.log("totallove per user = ", totallove, " from msg.sender = ", msg.sender);
-        console.log("trying to remove the visual with love: ", auction.visualNoVotes[visualId][category], " and totallove = ", auction.totalLove / 3);
+        console.log(
+            "trying to remove the visual with love: ",
+            auction.visualNoVotes[visualId][category],
+            " and totallove = ",
+            auction.totalLove / 3
+        );
 
-         if (auction.visualNoVotes[visualId][category] > auction.totalLove / 3) {
-             // a third of lovers has voted to remove the visual trait from the auction
-             uint k = 0;
+        if (auction.visualNoVotes[visualId][category] > auction.totalLove / 3) {
+            // a third of lovers has voted to remove the visual trait from the auction
+            uint256 k = 0;
             // identify the location of the visualId
-             for (uint i = 2; i < 10; i++) {
-                 // start with i=2 because we don't want people to remove the 2 parent's traits
-                 if(auction.visualIds[i][category] == visualId) {
+            for (uint256 i = 2; i < 10; i++) {
+                // start with i=2 because we don't want people to remove the 2 parent's traits
+                if (auction.visualIds[i][category] == visualId) {
                     console.log("EQUALITY");
                     k = i;
                     break;
                 }
             }
 
-             console.log("REMOVAL......... ", k);
+            console.log("REMOVAL......... ", k);
 
             require(k != 0, "The trait to be removed does not exist in the auction list");
 
-                uint j; 
-                // reset all values, so that new visuals can be submitted
-                for (j = k; j < 9 && auction.visualIds[j][category] != 0; j++) {  // stop at 9 because of the j+1 below
-                    auction.visualIds[j][category] = auction.visualIds[j+1][category];
-                    auction.visualIdVotes[j][category] = auction.visualIdVotes[j+1][category];
-                    auction.visualNoVotes[j][category] = auction.visualNoVotes[j+1][category];
-                } 
+            uint256 j;
+            // reset all values, so that new visuals can be submitted
+            for (j = k; j < 9 && auction.visualIds[j][category] != 0; j++) {
+                // stop at 9 because of the j+1 below
+                auction.visualIds[j][category] = auction.visualIds[j + 1][category];
+                auction.visualIdVotes[j][category] = auction.visualIdVotes[j + 1][category];
+                auction.visualNoVotes[j][category] = auction.visualNoVotes[j + 1][category];
+            }
 
-                console.log("stop removal at ... ", j);
+            console.log("stop removal at ... ", j);
 
-                 auction.visualIds[j+1][category] = 0;
-                 auction.visualIdVotes[j+1][category] = 0;
-                 auction.visualNoVotes[j+1][category] = 0;
-            
-         }
+            auction.visualIds[j + 1][category] = 0;
+            auction.visualIdVotes[j + 1][category] = 0;
+            auction.visualNoVotes[j + 1][category] = 0;
+        }
     }
 
     // TODO limits on when this can be called?
